src/requirements.txt
aioredis==1.1.0
amqp==2.3.2
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.7.1
Automat==0.7.0
Babel==2.6.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
channels==2.1.2
channels-redis==2.2.1
chardet==3.0.4
constantly==15.1.0
daphne==2.2.1
Django==2.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-redis==4.9.0
djangorestframework==3.8.2
flower==0.9.2
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kombu==4.2.1
Markdown==2.6.11
msgpack==0.5.6
oauthlib==2.1.0
psycopg2-binary==2.7.5
PyHamcrest==1.9.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
tornado==5.1
Twisted==18.7.0
txaio==18.7.1
urllib3==1.23
vine==1.1.4
zope.interface==4.5.0
